+++ api/libs/auth.js
@@ -46,7 +46,7 @@ const walletAuth = new WalletAuthenticator({
     name: env.appName,
     description: env.appDescription,
     icon,
-    link: isNetlify ? env.baseUrl.replace(netlifyPrefix, '') : env.baseUrl,
+    link: isNetlify ? env.baseUrl.replace(netlifyPrefix, '') : env.baseUrl.replace('3030', '3000'),
   },
   chainInfo: ({ locale }) => {
     if (locale === 'zh' && env.chainHostZh) {