+++ packages/did-playground/src/Action/actions.js
@@ -7,11 +7,11 @@ async function createSwapOrder(api) {
 }
 
 const getValidPayAmount = (payAmount, price) => {
-  if (payAmount !== undefined) {
+  if (Number(payAmount) > 0) {
     return payAmount;
   }
 
-  if (price !== undefined) {
+  if (Number(price) > 0) {
     return price;
   }
 
@@ -96,6 +96,7 @@ export const actions = {
       loc: getMessage(props.location, session),
       bg: props.backgroundUrl,
       logo: props.logoUrl,
+      svg: props.svg,
     }),
   },
   buy_foreign_ticket_with_local_token: {
@@ -175,6 +176,7 @@ export const actions = {
       loc: getMessage(props.location, session),
       bg: props.backgroundUrl,
       logo: props.logoUrl,
+      svg: props.svg,
     }),
   },
   buy_local_ticket_with_foreign_token: {
@@ -246,7 +248,7 @@ export const actions = {
     extraParams: (props, session) => ({
       pa: props.payAmount || 1,
       pt: 'certificate',
-      ra: props.receiveAmount || props.price || 1,
+      ra: getValidPayAmount(props.receiveAmount, props.price),
       rt: 'token',
       name: getMessage(props.name, session),
     }),
@@ -263,6 +265,7 @@ export const actions = {
       loc: getMessage(props.location, session),
       bg: props.backgroundUrl,
       logo: props.logoUrl,
+      svg: props.svg,
     }),
   },
   sell_local_badge_for_local_token: {
@@ -270,7 +273,7 @@ export const actions = {
     extraParams: (props, session) => ({
       pa: props.payAmount || 1,
       pt: 'badge',
-      ra: props.receiveAmount || props.price || 1,
+      ra: getValidPayAmount(props.receiveAmount, props.price),
       rt: 'token',
       name: getMessage(props.name, session),
     }),
@@ -294,7 +297,7 @@ export const actions = {
     extraParams: (props, session) => ({
       pa: props.payAmount || 1,
       pt: 'ticket',
-      ra: props.receiveAmount || props.price || 1,
+      ra: getValidPayAmount(props.receiveAmount, props.price),
       rt: 'token',
       name: getMessage(props.name, session),
     }),