+++ src/pages/profile.js
@@ -75,7 +75,6 @@ export default function ProfilePage() {
 }
 
 const Main = styled.main`
-  margin: 80px 0;
   display: flex;
 
   .avatar {